requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
requests
whitenoise
django-debug-toolbar
django-background-tasks